requirements.txt
amqp==5.0.9
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==2.2
django-cors-headers==3.0.0
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
djangorestframework==3.12.4
et-xmlfile==1.1.0
greenlet==1.1.2
html5lib==1.1
importlib-metadata==4.8.2
Jinja2==3.0.3
kombu==5.2.3
ldap3==2.9.1
lxml==4.7.1
MarkupSafe==2.0.1
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
prettytable==2.4.0
prompt-toolkit==3.0.24
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyecharts==1.7.0
pymssql==2.2.3
pyodbc==4.0.32
python-dateutil==2.8.2
python-ldap @ file:///C:/inetpub/wwwroot/Visualization/python_ldap-3.3.1-cp37-cp37m-win_amd64.whl
pytz==2021.3
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.2
typing-extensions==4.0.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wfastcgi==3.0.0
xlwings==0.25.3
zipp==3.6.0
